+++ framework/LoopMeUnitedSDK/AdConfiguration/AdConfiguration/VAST/VastProperties.swift
@@ -250,19 +250,26 @@ extension VastProperties: XMLParserDelegate {
     func parseMediaFiles() throws {
         try tempMediaFiles.sort(by: sortMediaFilesBlock)
         var videoURLs: [String] = []
+        var video360URLs: [String] = []
         for mediaFileNode in tempMediaFiles {
             let delivery = mediaFileNode.props["delivery"]
             let framework = mediaFileNode.props["apiFramework"]
             let type = mediaFileNode.props["type"]
+            let mediaType = mediaFileNode.props["mediaType"]
             if let content = mediaFileNode.content {
                 if delivery == "progressive" && type == "video/mp4" {
-                    videoURLs.append(content.trimmingCharacters(in: .whitespacesAndNewlines))
+                    if mediaType == "360" {
+                        video360URLs.append(content.trimmingCharacters(in: .whitespacesAndNewlines))
+                    } else {
+                        videoURLs.append(content.trimmingCharacters(in: .whitespacesAndNewlines))
+                    }
                 } else if framework == "VPAID" && type == "application/javascript" {
                     assetLinks.vpaidURL = content.trimmingCharacters(in: .whitespacesAndNewlines)
                 }
             }
         }
         assetLinks.videoURL = videoURLs
+        assetLinks.video360URL = video360URLs
     }
     
     func sortMediaFilesBlock(node1: Node, node2: Node) throws -> Bool {