+++ framework/LoopMeUnitedSDK/Core/Viewability/OMSDK/OMSDK_Loopme.framework/Versions/A/Headers/OMIDVASTProperties.h
@@ -8,7 +8,7 @@
 #import <UIKit/UIKit.h>
 
 /**
- *  List of supported video player positions.
+ *  List of supported media player positions.
  */
 typedef NS_ENUM(NSUInteger, OMIDPosition) {
     OMIDPositionPreroll,
@@ -28,11 +28,11 @@ typedef NS_ENUM(NSUInteger, OMIDPosition) {
 @property(nonatomic, readonly) OMIDPosition position;
 
 /**
- *  This method enables the video player to create a new VAST properties instance for skippable video ad placement.
+ *  This method enables the media player to create a new VAST properties instance for skippable media ad placement.
  *
  * @param skipOffset The number of seconds before the skip button is presented.
- * @param autoPlay Determines whether the video will auto-play content.
- * @param position The position of the video in relation to other content.
+ * @param autoPlay Determines whether the media will auto-play content.
+ * @param position The position of the media in relation to other content.
  * @return A new instance of VAST properties.
  */
 - (nonnull instancetype)initWithSkipOffset:(CGFloat)skipOffset
@@ -40,15 +40,20 @@ typedef NS_ENUM(NSUInteger, OMIDPosition) {
                                   position:(OMIDPosition)position;
 
 /**
- *  This method enables the video player to create a new VAST properties instance for non-skippable video ad placement.
+ *  This method enables the media player to create a new VAST properties instance for non-skippable media ad placement.
  *
- * @param autoPlay Determines whether the video will auto-play content.
- * @param position The position of the video in relation to other content.
+ * @param autoPlay Determines whether the media will auto-play content.
+ * @param position The position of the media in relation to other content.
  * @return A new instance of VAST properties.
  */
 - (nonnull instancetype)initWithAutoPlay:(BOOL)autoPlay
                                 position:(OMIDPosition)position;
 
 - (null_unspecified instancetype)init NS_UNAVAILABLE;
+
+/**
+ * This method should not be called by the integration
+ */
+- (NSDictionary *_Nonnull)toJSON;
 
 @end